requirements.txt
pandas
nltk
joblib
textblob
numpy
scikit-learn
imbalanced-learn
opencv-python
pytesseract
Pillow
flask
pyspellchecker
re